backend/requirements.txt
# IntelliGuard Enterprise Requirements
# Production-grade dependencies with security and performance optimizations

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching & Session Management
redis==5.0.1
hiredis==2.2.3

# Machine Learning & Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography==41.0.8

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Monitoring & Metrics
prometheus-client==0.19.0
psutil==5.9.6

# Logging & Observability
structlog==23.2.0
python-json-logger==2.0.7

# Email & Notifications
aiosmtplib==3.0.1
jinja2==3.1.2

# File Processing
python-magic==0.4.27
openpyxl==3.1.2

# Validation & Parsing
validators==0.22.0
python-dateutil==2.8.2

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Performance
orjson==3.9.10
ujson==5.8.0

# Deployment
gunicorn==21.2.0
supervisor==4.2.5

# Development Tools
ipython==8.17.2
jupyter==1.0.0

# Additional Security
bandit==1.7.5
safety==2.3.5